requirements.txt
matplotlib
numpy
pillow
opencv-python
scikit-learn
flask
flask-cors
xgboost
seaborn
pandas
